requirements.txt
Flask
Flask-SQLAlchemy
mysql-connector-python
PyJWT
python-dotenv
argon2-cffi
sqlalchemy